+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -120,7 +120,7 @@ export type DraftEditorSettings = Pick<
 export interface MediaSettings {
   pauseMedia?: boolean;
   disableRightClick?: boolean;
-  onMediaUpload?: (isUploadingMedia: boolean) => void;
+  onUpload?: (isUploading: boolean) => void;
 }
 
 export interface LinkSettings {